Initial refactor.  Everything still Bdb but storage component is now split out.
