+++ scalaengine/src/main/scala/server/StorageHandler.scala
@@ -134,11 +134,11 @@ class StorageHandler(env: Environment, val root: ZooKeeperProxy#ZooKeeperNode, v
    *   (3) valueSchema is already set in the namespace/valueSchema file
    *       in ZooKeeper
    */
-  private def makePartitionHandler(
+  private def makeBdbPartitionHandler(
       database: Database, namespace: String, partitionIdLock: ZooKeeperProxy#ZooKeeperNode,
       startKey: Option[Array[Byte]], endKey: Option[Array[Byte]]) = {
     val schemas = schemasFor(namespace)
-    new PartitionHandler(database, partitionIdLock, startKey, endKey, getNamespaceRoot(namespace), schemas._1, schemas._2)
+    new PartitionHandler(new BdbStorageManager(database, partitionIdLock, startKey, endKey, getNamespaceRoot(namespace), schemas._1, schemas._2))
   }
 
   /** Iterator scans the entire cursor and does not close it */
@@ -208,7 +208,7 @@ class StorageHandler(env: Environment, val root: ZooKeeperProxy#ZooKeeperNode, v
 
       /* Make partition handler */
       val db      = makeDatabase(request.namespace, keySchemaFor(request.namespace), None)
-      val handler = makePartitionHandler(db, request.namespace, partitionIdLock, request.startKey, request.endKey)
+      val handler = makeBdbPartitionHandler(db, request.namespace, partitionIdLock, request.startKey, request.endKey)
 			//val handler = makePartitionHandlerWithAC(db, acdb, request.namespace, partitionIdLock, request.startKey, request.endKey)
 
       /* Add to our list of open partitions */
@@ -313,7 +313,7 @@ class StorageHandler(env: Environment, val root: ZooKeeperProxy#ZooKeeperNode, v
           txn.commit()
 
           /* Make partition handler from request */
-          val handler = makePartitionHandler(newDb, namespace, partitionIdLock, startKey, endKey)
+          val handler = makeBdbPartitionHandler(newDb, namespace, partitionIdLock, startKey, endKey)
 					//val handler = makePartitionHandlerWithAC(newDb, acDb, namespace, partitionIdLock, startKey, endKey)
 
           /* Add to our list of open partitions */
@@ -337,63 +337,77 @@ class StorageHandler(env: Environment, val root: ZooKeeperProxy#ZooKeeperNode, v
 
         /* Get the handler and shut it down */
         val handler = Option(partitions.remove(partitionId)) getOrElse {reply(InvalidPartition(partitionId)); return}
-
-        val dbName = handler.db.getDatabaseName /* dbName is namespace */
-        val dbEnv  = handler.db.getEnvironment
-
-        //logger.info("Unregistering handler from MessageHandler for partition %s in namespace %s".format(partitionId, dbName))
         handler.stopListening
 
-        logger.info("[%s] Deleting partition [%s, %s) for namespace %s", this, JArrays.toString(handler.startKey.orNull), JArrays.toString(handler.endKey.orNull), dbName)
+        handler.manager match {
+          case bdbManager:BdbStorageManager => {
+            val dbName = bdbManager.db.getDatabaseName /* dbName is namespace */
+            val dbEnv  = bdbManager.db.getEnvironment
 
-        val ctx = getContextForNamespace(dbName) getOrElse {
-          /**
-           * Race condition - a request to delete a namespace is going on
-           * right now- this error can actually be safely ignored since this
-           * partition will be deleted (in response to the namespace deletion
-           * request)
-           */
-          reply(RequestRejected("Partition will be removed by a delete namespace request currently in progress", msg)); return
-        }
+            //logger.info("Unregistering handler from MessageHandler for partition %s in namespace %s".format(partitionId, dbName))
 
-        ctx.synchronized {
-          val succ = ctx.partitions.remove((partitionId, handler))
-          assert(succ, "Handler not successfully removed from partitions")
+            logger.info("[%s] Deleting partition [%s, %s) for namespace %s", this, JArrays.toString(bdbManager.startKey.orNull), JArrays.toString(bdbManager.endKey.orNull), dbName)
 
-          /* Delete from partitionDB, and (possibly) delete the database in a
-          * single transaction */
-          val txn = env.beginTransaction(null, null)
-
-          /* Remove from bdb map */
-          partitionDb.delete(txn, new DatabaseEntry(partitionId.getBytes))
-
-          if (ctx.partitions.isEmpty) {
-            /* Remove database from environment- this removes all the data
-            * associated with the database */
-            logger.info("[%s] Deleting database %s for namespace %s".format(this, dbName, dbName))
-            handler.db.close() /* Close underlying DB */
-            dbEnv.removeDatabase(txn, dbName)
-          } else {
-            logger.info("[%s] Garbage collecting inaccessible keys, since %d partitions in namespace %s remain", this, ctx.partitions.size, dbName)
-
-            implicit def orderedByteArrayView(thiz: Array[Byte]) = new Ordered[Array[Byte]] {
-              def compare(that: Array[Byte]) = ctx.comparator.compare(thiz, that) 
+            val ctx = getContextForNamespace(dbName) getOrElse {
+              /**
+               * Race condition - a request to delete a namespace is going on
+               * right now- this error can actually be safely ignored since this
+               * partition will be deleted (in response to the namespace deletion
+               * request)
+               */
+              reply(RequestRejected("Partition will be removed by a delete namespace request currently in progress", msg)); return
             }
 
-            val thisPartition = handler
-            var thisSlice = Slice(thisPartition.startKey, thisPartition.endKey)
-            ctx.partitions.foreach { t =>
-              val thatPartition = t._2
-              val thatSlice = Slice(thatPartition.startKey, thatPartition.endKey)
-              thisSlice = thisSlice.remove(thatSlice) /* Remove (from deletion) slice which cannot be deleted */
+            ctx.synchronized {
+              val succ = ctx.partitions.remove((partitionId, handler))
+              assert(succ, "Handler not successfully removed from partitions")
+
+              /* Delete from partitionDB, and (possibly) delete the database in a
+               * single transaction */
+              val txn = env.beginTransaction(null, null)
+
+              /* Remove from bdb map */
+              partitionDb.delete(txn, new DatabaseEntry(partitionId.getBytes))
+
+              if (ctx.partitions.isEmpty) {
+                /* Remove database from environment- this removes all the data
+                 * associated with the database */
+                logger.info("[%s] Deleting database %s for namespace %s".format(this, dbName, dbName))
+                bdbManager.db.close() /* Close underlying DB */
+                dbEnv.removeDatabase(txn, dbName)
+              } else {
+                logger.info("[%s] Garbage collecting inaccessible keys, since %d partitions in namespace %s remain", this, ctx.partitions.size, dbName)
+
+                implicit def orderedByteArrayView(thiz: Array[Byte]) = new Ordered[Array[Byte]] {
+                  def compare(that: Array[Byte]) = ctx.comparator.compare(thiz, that) 
+                }
+
+                val thisPartition = bdbManager
+                var thisSlice = Slice(thisPartition.startKey, thisPartition.endKey)
+                ctx.partitions.foreach { t =>
+                  t._2 match {
+                    case thatPartition:BdbStorageManager => {                   
+                      val thatSlice = Slice(thatPartition.startKey, thatPartition.endKey)
+                      thisSlice = thisSlice.remove(thatSlice) /* Remove (from deletion) slice which cannot be deleted */
+                    }
+                    case _ =>
+                      logger.warning("Can't properly garbage collect when namespace mixes Bdb and other partitions")
+                  }
+                }
+                thisSlice.foreach((startKey, endKey) => {
+                  logger.info("++ [%s] Deleting range: [%s, %s)", this, JArrays.toString(startKey.orNull), JArrays.toString(endKey.orNull))
+                  bdbManager.deleteRange(startKey, endKey, txn)
+                })
+              }
+
+              txn.commit()
             }
-            thisSlice.foreach((startKey, endKey) => {
-              logger.info("++ [%s] Deleting range: [%s, %s)", this, JArrays.toString(startKey.orNull), JArrays.toString(endKey.orNull))
-              handler.deleteRange(startKey, endKey, txn)
-            })
           }
-
-          txn.commit()
+          case _ => {
+            handler.stop
+            reply(RequestRejected("Unknown StorageManager type.  Cannot delete partition", msg))
+            return
+          }
         }
 
         handler.stop
